Rename project_1.py to joecalculatorproject.py
Goal:
Create a calculator that can solve simple math equations involving addition, subtraction, multiplication, and division using the Python scripting language. 
How to Run:
Start the program using an interpreter/compiler for python
When the program opens, you will be asked to choose a number ranging from 1 to 4, representing a mathematical operation that will be used for your problem. Afterward, you will be asked to enter a first and second number. 
The calculator will then solve your equation using the operation that you selected
The program will then display the answer. When you are ready to close the program, press the enter button and the program will close itself.