Update joecalculatorproject.py
Joe Belcher
IT 3038C

Project 1: Python Calculator

Goal:
Create a calculator that can solve simple math equations involving addition, subtraction, multiplication, and division using the Python scripting language. 
How to Run:
Start the program using an interpreter/compiler for python
When the program opens, you will be asked to choose a number ranging from 1 to 4, representing a mathematical operation that will be used for your problem. Afterward, you will be asked to enter a first and second number. 
The calculator will then solve your equation using the operation that you selected
The program will then display the answer. When you are ready to close the program, press the enter button and the program will close itself. 
Citations:
https://www.digitalocean.com/community/tutorials/how-to-make-a-calculator-program-in-python-3
The core of the calculator uses the Python code found here, I changed the input options for the operators to numbers and made sure that program did not automatically close itself upon completion. I also made sure to add a unique message for an improper mathematical operator and for closing the program using the enter button. 
My future additions would be to add more options for numbers in the equation beyond two at once, as well as adding more operations that the calculator can perform.